CHROMIUM: iwl7000: mvm: add missing lq_color
In the compressed BA notif, the driver didn't parse out
the LQ color, so statistics for the rates tried were
always thrown out. Add it so it gets correctly used.

While at it, fix the name of the relevant field in the
struct.

Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
iwl7000-tree: 6ba187969edeaa8b428653326290f70bd70b51ea