CHROMIUM: iwl7000: mac80211: remove BUG() when interface type is invalid
In the ieee80211_setup_sdata() we check if the interface type is valid
and, if not, call BUG().  This should never happen, but if there is
something wrong with the code, it will not be caught until the bug
happens when an interface is being set up.  Calling BUG() is too
extreme for this and a WARN_ON() would be better used instead.  Change
that.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 31ac09f78ed634aa641adc88faad685f582c1393