CHROMIUM: kaiser: setup_pcid() beware in setting X86_CR4_PCIDE
3.8's set_in_cr4() includes side-effects on mmu_cr4_features and
trampoline_cr4_features, which prevent the secondary processors
from booting successfully: and so then the machine immediately
reboots (unless "nopcid" or PCID unsupported).

4.0's commit 375074cc736a ("x86: Clean up cr4 manipulation") is a
more extensive patch, which introduced cr4_set_bits() to fix this
properly (and 3.18.72 already contained a backport of that): but
as usual I'm torn between last-minute backporting a larger commit,
that might introduce other dependencies, or simply fixing the boot.

Expediency chooses the latter option: while accepting that a proper
backport may turn out to be needed later (in particular, is it right
or wrong not to update mmu_cr4_features? hibernation testing may show).
4.14's commit c7ad5ad297e6 ("x86/mm/64: Initialize CR4.PCIDE early")
would be another to consider if choosing to backport more here.

Change-Id: I8920fba55dd47c2fc177b471dd76f5ea9d43ae7b
Signed-off-by: Jorge Lucangeli Obes <jorgelo@chromium.org>