CHROMIUM: iwl7000: mvm: fix "failed to remove key" message
When the GTK is installed, we install it to HW with the
station ID of the AP.

Mac80211 will try to remove it only after the AP sta is
removed, which will result in a failure to remove key
since we do not have any station for it.

This is a valid situation, but a previous commit removed
the early return and added a return with error value, which
resulted in an error message that is confusing to users.

Remove the error return value.

BUG=b:72754823
TEST=Connect and disconnect from a protected AP.  Check that the
     "failed to remove key (0, ff:ff:ff:ff:ff:ff) from hardware (-22)"
     message doesn't appear in the logs.

Change-Id: I7e23ef13e3206869bd6e70808c6d824fbedbed55
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: cd0d2246a36f4e172808a7c51b7c8166a516dded
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/962001
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>