CHROMIUM: iwl7000: mvm: Move unused phy's to a default channel
When immediate quiet bit is set in CSA, the entire channel is blocked
by the firmware. It is expected that all the MACs will evacuate the
channel and the phy will be eventually either moved or removed.
Currently, the phy context is just unreferenced and thus, the quiet bit
is kept set and it will be impossible to TX on this phy, if we will need
to reuse it in the future.
This can be seen when doing a channel switch with mode=1 (quiet) twice
from channel X to Y and then back to channel X.
Fix that, by moving the phy context to a default channel when not
referenced anymore.

Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
automatic-review: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 4bc8af12dcd45aa05730d3cc5b090f8515123e6c