CHROMIUM: iwl7000: pcie: support host cmds in vendor mode
Two mentionable changes in the vendor mode cmd sending flow are:
First, the use of a new tx transfer descriptor, instead
of the legacy tfd.
Second, the vendor mode separates completely between the
tx flow for sending the command, to the rx flow in which
you receive a notification with the data the hcmd asked for.

The new flow is as depicted:
For every command sent by the host to the fw, a completion
is sent to the host in a sense of an interrupt, with no rx
buffer attached. The completion refers to the queue the command
was sent from, and the transfer descriptor id.
Later, if the command asked from the fw for a notification,
a notification would be sent.
This breaks the current sync commands implementation.

In this patch, do:
- Support sending hcmds complying to the vendor mode.
- Introduce the new tx transfer descriptor, and use it,
  instead of the legacy TFD.
- Add a mechanism for allowing sync commands, to avoid
  breaking the op mode layer.

restriction=nopublic

Signed-off-by: Golan Ben Ami <golan.ben.ami@intel.com>
iwl7000-tree: 65e23f20c0c9a40073bc453439ec49920a7aad1f