CHROMIUM: config: Reenable iwl7000 without needing config fragments
This effectively reverts
commit 189058544cc0 ("CHROMIUM: config: renormalize config after iwl7000 changes")
which had removed iwl7000 and friends because it relied on !cfg80211.

This CL was done with
`echo CONFIG_IWL7000=m >> chromeos/config/x86_64/chromeos-intel-pineview.flavour.config`
followed by
`./chromeos/scripts/kernelconfig olddefconfig`

The Core31 version of the iwl7000 brought in an iwl7000-specific cfg80211
module (iwl7000_cfg80211.ko) which could not co-exist with cfg80211.ko.
These changes were made to enable a setup where only one of Intel-wifi or
non-Intel-wifi stacks was enabled on a given board image.
With Core35, we revert that, and go back to Intel wifi using stock cfg80211,
and hence the need to enable it in our configs and stop using the
"intelwifi38" config-fragment.

The following changes should land together to enable this transition
in a single step:
- CL:1006658 (add Core35 update to iwl7000)
- CL:1008644 (revert the "intelwifi38" USE-flag based config fragment since iwlwifi is enabled by default now)
- CL:1008844 (revert tricky's use of "intelwifi38")
- CL:*606968 (revert falco_li's use of "intelwifi38")
- CL:1008842 [this CL] (add iwl7000 back to the kernel config and renormalize)

CQ-DEPEND=CL:1006658, CL:1008644, CL:1008844, CL:*606968
BUG=b:76340244
TEST=build on falco_li, link

Signed-off-by: Kirtika Ruchandani <kirtika@google.com>
Change-Id: I4d292fa1377e7e105c333951cc565c0135992cab
Reviewed-on: https://chromium-review.googlesource.com/1008842
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Rajat Jain <rajatja@chromium.org>